prefer one time move commands over window rules in wayland sway/hyprland (#77)
When using ueberzug with vifm, we sometimes want to show an image on
either the left panel or the right panel. Each time the location of the
image is changed, a window rule is emitted, which caused inconsistent
rules. This caused the location of the image sometimes to be wrong in
sway (specifically when trying to preview an image once in the left
panel, then the right side panel, then the left side panel again).
It also doesn't feel right to use window rules at all, and I would
prefer to always use one time commands instead, since window rules are
remembered by the window manager, which isn't desirable.